+++ ECMA-SL/semantics/core/functorial/interpreter_functor_intf.ml
@@ -73,7 +73,6 @@ module type P = sig
     val add_memory : t -> memory -> t
     val add_func : t -> string -> Func.t -> unit
   end
-
 end
 
 module type S = sig