+++ ECMA-SL/syntax/extended/eExpr.ml
@@ -30,8 +30,8 @@ let float_str (f : float) : string =
   if is_special_number f_str || String.contains f_str '.' then f_str
   else f_str ^ ".0"
 
-let pp_custom_val (pp_inner_val : Fmt.t -> Smtml.Value.t -> unit) (ppf : Fmt.t) (v : Smtml.Value.t) :
-  unit =
+let pp_custom_val (pp_inner_val : Fmt.t -> Smtml.Value.t -> unit) (ppf : Fmt.t)
+  (v : Smtml.Value.t) : unit =
   let open Fmt in
   match v with
   | App (`Op "null", []) -> format ppf "null"
@@ -41,18 +41,18 @@ let pp_custom_val (pp_inner_val : Fmt.t -> Smtml.Value.t -> unit) (ppf : Fmt.t) 
   | Str s -> format ppf "%S" s
   | True -> format ppf "true"
   | False -> format ppf "false"
-  | App (`Op "symbol", [Str s]) -> format ppf "'%s" s
-  | App (`Op "loc", [Int l]) -> Loc.pp ppf l
+  | App (`Op "symbol", [ Str s ]) -> format ppf "'%s" s
+  | App (`Op "loc", [ Int l ]) -> Loc.pp ppf l
   | List lst -> format ppf "[%a]" (pp_lst !>", " pp_inner_val) lst
-  | App (`Op "NullType" , []) -> format ppf "null" 
-  | App (`Op "IntType" , []) -> format ppf "int" 
-  | App (`Op "FltType" , []) -> format ppf "float" 
-  | App (`Op "StrType" , []) -> format ppf "string" 
-  | App (`Op "BoolType" , []) -> format ppf "bool" 
-  | App (`Op "SymbolType" , []) -> format ppf "symbol" 
-  | App (`Op "LocType" , []) -> format ppf "object" 
-  | App (`Op "ListType" , []) -> format ppf "list" 
-  | App (`Op "CurryType" , []) -> format ppf "curry" 
+  | App (`Op "NullType", []) -> format ppf "null"
+  | App (`Op "IntType", []) -> format ppf "int"
+  | App (`Op "FltType", []) -> format ppf "float"
+  | App (`Op "StrType", []) -> format ppf "string"
+  | App (`Op "BoolType", []) -> format ppf "bool"
+  | App (`Op "SymbolType", []) -> format ppf "symbol"
+  | App (`Op "LocType", []) -> format ppf "object"
+  | App (`Op "ListType", []) -> format ppf "list"
+  | App (`Op "CurryType", []) -> format ppf "curry"
   | App (`Op fn, fvs) ->
     format ppf "{%S}@(%a)" fn (pp_lst !>", " pp_inner_val) fvs
   | _ -> Log.fail "Val.pp_custom_val: unexpected case"
@@ -83,7 +83,9 @@ let rec pp (ppf : Fmt.t) (e : t) : unit =
   | Symbolic (t, e') -> format ppf "se_mk_symbolic(%a, %a)" Type.pp t pp e'
 
 let str (e : t) : string = Fmt.str "%a" pp e
-let isvoid (e : t) : bool = match e.it with Val App (`Op "void", []) -> true | _ -> false
+
+let isvoid (e : t) : bool =
+  match e.it with Val (App (`Op "void", [])) -> true | _ -> false
 
 let rec map (mapper : t -> t) (e : t) : t =
   let map' = map mapper in